+++ test/src/analyzers/lint_analyzer/rules/rules_list/member_ordering/examples/not_widget_example.dart
@@ -1,0 +1,23 @@
+class MyNotWidget {
+  const MyNotWidget();
+
+  void doNothing() {}
+
+  @override
+  void build(Object context) {}
+
+  @override
+  void initState() {}
+
+  @override
+  void didChangeDependencies() {}
+
+  @override
+  void didUpdateWidget() {}
+
+  @override
+  void dispose() {}
+
+  @override
+  void someOtherMethod() {}
+}