+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * feat: support report to the json file option for the `analyze` command.
 * feat: make CliRunner a part of public API in order to support transitive executable calls use-case.
 * feat: add static code diagnostic [`avoid-cascade-after-if-null`](https://dartcodemetrics.dev/docs/rules/common/avoid-cascade-after-if-null).
+* feat: **Breaking change** handle widget members order separately in [`member-ordering`](https://dartcodemetrics.dev/docs/rules/common/member-ordering).
 
 ## 4.21.2